+++ soul-admin/src/main/java/org/dromara/soul/admin/config/NacosConfiguration.java
@@ -20,6 +20,7 @@ package org.dromara.soul.admin.config;
 import com.alibaba.nacos.api.NacosFactory;
 import com.alibaba.nacos.api.PropertyKeyConst;
 import com.alibaba.nacos.api.config.ConfigService;
+import org.dromara.soul.admin.config.properties.NacosProperties;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
 import org.springframework.context.annotation.Bean;
