+++ soul-admin/src/main/java/org/dromara/soul/admin/config/ZookeeperConfiguration.java
@@ -18,6 +18,7 @@
 package org.dromara.soul.admin.config;
 
 import org.I0Itec.zkclient.ZkClient;
+import org.dromara.soul.admin.config.properties.ZookeeperProperties;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
 import org.springframework.context.annotation.Bean;
