+++ soul-admin/src/main/java/org/dromara/soul/admin/config/properties/WebsocketSyncProperties.java
@@ -15,31 +15,25 @@
  * limitations under the License.
  */
 
-package org.dromara.soul.admin.config;
+package org.dromara.soul.admin.config.properties;
 
 import lombok.Getter;
 import lombok.Setter;
 import org.springframework.boot.context.properties.ConfigurationProperties;
 
-import java.time.Duration;
-
 /**
- * the http sync strategy properties.
- * @author huangxiaofeng
+ * the websocket sync strategy properties.
+ *
+ * @author xiaoyu
  */
 @Getter
 @Setter
-@ConfigurationProperties(prefix = "soul.sync.http")
-public class HttpSyncProperties {
+@ConfigurationProperties(prefix = "soul.sync.websocket")
+public class WebsocketSyncProperties {
 
     /**
-     * Whether enabled http sync strategy, default: true.
+     * default: true.
      */
     private boolean enabled = true;
-
-    /**
-     * Periodically refresh the config data interval from the database, default: 5 minutes.
-     */
-    private Duration refreshInterval = Duration.ofMinutes(5);
 
 }
