+++ soul-admin/src/main/java/org/dromara/soul/admin/controller/DashboardUserController.java
@@ -19,7 +19,7 @@ package org.dromara.soul.admin.controller;
 
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.collections4.CollectionUtils;
-import org.dromara.soul.admin.config.SecretProperties;
+import org.dromara.soul.admin.config.properties.SecretProperties;
 import org.dromara.soul.admin.dto.DashboardUserDTO;
 import org.dromara.soul.admin.page.CommonPager;
 import org.dromara.soul.admin.page.PageParameter;
