+++ soul-admin/src/main/java/org/dromara/soul/admin/controller/IndexController.java
@@ -17,7 +17,7 @@
 
 package org.dromara.soul.admin.controller;
 
-import org.dromara.soul.admin.listener.SoulDomain;
+import org.dromara.soul.admin.utils.SoulDomain;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -42,5 +42,4 @@ public class IndexController {
         model.addAttribute("domain", SoulDomain.getInstance().getHttpPath());
         return "index";
     }
-
 }
