+++ soul-admin/src/main/java/org/dromara/soul/admin/controller/MetaDataController.java
@@ -145,8 +145,7 @@ public class MetaDataController {
         }
         return SoulAdminResult.success(SoulResultMessage.ENABLE_SUCCESS);
     }
-
-
+    
     /**
      * Sync metadata.
      *
