+++ soul-admin/src/main/java/org/dromara/soul/admin/mapper/PermissionMapper.java
@@ -78,8 +78,7 @@ public interface PermissionMapper {
      * @return rows int
      */
     int deleteByObjectIdAndResourceId(PermissionQuery permissionQuery);
-
-
+    
     /**
      * delete permission by object id.
      *
