+++ soul-admin/src/main/java/org/dromara/soul/admin/service/impl/AppAuthServiceImpl.java
@@ -378,5 +378,4 @@ public class AppAuthServiceImpl implements AppAuthService {
         }
         return data;
     }
-
 }
