+++ soul-admin/src/main/java/org/dromara/soul/admin/service/init/LocalDataSourceLoader.java
@@ -22,7 +22,7 @@ import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.lang3.StringUtils;
 import org.apache.ibatis.io.Resources;
 import org.apache.ibatis.jdbc.ScriptRunner;
-import org.dromara.soul.admin.config.DataBaseProperties;
+import org.dromara.soul.admin.config.properties.DataBaseProperties;
 import org.springframework.beans.BeansException;
 import org.springframework.beans.factory.config.InstantiationAwareBeanPostProcessor;
 import org.springframework.boot.autoconfigure.jdbc.DataSourceProperties;
