+++ soul-admin/src/test/java/org/dromara/soul/admin/config/HttpSyncPropertiesTest.java
@@ -18,6 +18,7 @@
 package org.dromara.soul.admin.config;
 
 import org.dromara.soul.admin.AbstractConfigurationTest;
+import org.dromara.soul.admin.config.properties.HttpSyncProperties;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.mockito.junit.MockitoJUnitRunner;
