+++ soul-admin/src/test/java/org/dromara/soul/admin/config/WebsocketSyncPropertiesTest.java
@@ -18,6 +18,7 @@
 package org.dromara.soul.admin.config;
 
 import org.dromara.soul.admin.AbstractConfigurationTest;
+import org.dromara.soul.admin.config.properties.WebsocketSyncProperties;
 import org.junit.Test;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
 import org.springframework.context.annotation.Configuration;
